+++ packages/@expo/config-plugins/src/ios/__tests__/RequiresFullScreen-test.ts
@@ -1,15 +1,8 @@
 import * as WarningAggregator from '../../utils/warnings';
 import { getRequiresFullScreen, setRequiresFullScreen } from '../RequiresFullScreen';
 
-it(`returns true if nothing provided`, () => {
-  expect(getRequiresFullScreen({ ios: {} })).toBe(true);
-});
-it(`returns false if nothing provided after 43`, () => {
-  expect(getRequiresFullScreen({ ios: {}, sdkVersion: '43.0.0' })).toBe(false);
-  expect(getRequiresFullScreen({ ios: {}, sdkVersion: 'UNVERSIONED' })).toBe(false);
-});
-it(`asserts invalid SDK version`, () => {
-  expect(() => getRequiresFullScreen({ ios: {}, sdkVersion: '43.0' })).toThrow(/version/);
+it(`defaults to false`, () => {
+  expect(getRequiresFullScreen({ ios: {} })).toBe(false);
 });
 
 it(`returns the given value if provided`, () => {
@@ -18,12 +11,12 @@ it(`returns the given value if provided`, () => {
 });
 
 it(`sets UIRequiresFullScreen value`, () => {
-  expect(setRequiresFullScreen({ ios: { requireFullScreen: false } }, {})).toMatchObject({
-    UIRequiresFullScreen: false,
+  expect(setRequiresFullScreen({ ios: { requireFullScreen: true } }, {})).toMatchObject({
+    UIRequiresFullScreen: true,
   });
 
   expect(setRequiresFullScreen({}, {})).toMatchObject({
-    UIRequiresFullScreen: true,
+    UIRequiresFullScreen: false,
   });
 });
 
@@ -33,8 +26,18 @@ beforeEach(() => {
   WarningAggregator.addWarningIOS = jest.fn();
 });
 
+it(`sets requires full screen`, () => {
+  expect(setRequiresFullScreen({ ios: { requireFullScreen: true } }, {})).toMatchObject({
+    UIRequiresFullScreen: true,
+  });
+
+  expect(WarningAggregator.addWarningIOS).not.toBeCalled();
+});
+
 it(`updates the UISupportedInterfaceOrientations~ipad values to support iPad multi-tasking`, () => {
-  expect(setRequiresFullScreen({ ios: { requireFullScreen: false } }, {})).toMatchObject({
+  expect(
+    setRequiresFullScreen({ ios: { requireFullScreen: false, supportsTablet: true } }, {})
+  ).toMatchObject({
     UIRequiresFullScreen: false,
     'UISupportedInterfaceOrientations~ipad': [
       'UIInterfaceOrientationPortrait',
@@ -47,7 +50,17 @@ it(`updates the UISupportedInterfaceOrientations~ipad values to support iPad mul
   expect(WarningAggregator.addWarningIOS).not.toBeCalled();
 });
 
-it(`warns when the predefined UISupportedInterfaceOrientations~ipad values are invalid`, () => {
+it(`does not set the UISupportedInterfaceOrientations~ipad values to support iPad multi-tasking if supportsTablet is false`, () => {
+  expect(
+    setRequiresFullScreen({ ios: { requireFullScreen: false, supportsTablet: false } }, {})
+  ).toMatchObject({
+    UIRequiresFullScreen: false,
+  });
+
+  expect(WarningAggregator.addWarningIOS).not.toBeCalled();
+});
+
+xit(`warns when the predefined UISupportedInterfaceOrientations~ipad values are invalid`, () => {
   setRequiresFullScreen(
     { ios: { requireFullScreen: false } },
     {
@@ -91,9 +104,9 @@ it(`does not warn when predefined orientation mask is empty`, () => {
 
 it(`cannot remove predefined orientation mask`, () => {
   let plist = {};
-  plist = setRequiresFullScreen({ ios: { requireFullScreen: false } }, plist);
+  plist = setRequiresFullScreen({ ios: { requireFullScreen: false, supportsTablet: true } }, plist);
   expect(plist['UISupportedInterfaceOrientations~ipad']?.length).toBe(4);
-  plist = setRequiresFullScreen({ ios: { requireFullScreen: true } }, plist);
+  plist = setRequiresFullScreen({ ios: { requireFullScreen: true, isTabletOnly: true } }, plist);
   expect(plist['UISupportedInterfaceOrientations~ipad']?.length).toBe(4);
   expect(WarningAggregator.addWarningIOS).not.toBeCalled();
 });